Added German Translation
I really love this editor and so I want to support as well bring forward
this lovely editor!
So I made a translation for German.
I now have an enconding issue with the following letters:
Ã¶Ã¤Ã¼ÃÃÃÃ
I don't find a way to fix this issue, I holp someone can help me ;)